FuScript Integration for BBEdit on MacOS
This project has three AppleScripts that connect BBEdit on MacOS to the Fusion Studio via the FuScript command line process. This will free you from the constraints of the single line entry text field in the Fusion Console tab.

Fusion Studio needs to be running for the FuScript tool to function correctly. The fusion:, fu:, app:, and comp: pointers are configured and the active Fusion comp is linked to automatically. This means you are instantly capable of connecting to the foreground Fusion composite or able to create a Fusion UI Manger based script GUI.